+++ packages/core-js/modules/es.array.every.js
@@ -2,10 +2,14 @@
 var $ = require('../internals/export');
 var $every = require('../internals/array-iteration').every;
 var sloppyArrayMethod = require('../internals/sloppy-array-method');
+var arrayMethodUsesToLength = require('../internals/array-method-uses-to-length');
+
+var SLOPPY_METHOD = sloppyArrayMethod('every');
+var USES_TO_LENGTH = arrayMethodUsesToLength('every');
 
 // `Array.prototype.every` method
 // https://tc39.github.io/ecma262/#sec-array.prototype.every
-$({ target: 'Array', proto: true, forced: sloppyArrayMethod('every') }, {
+$({ target: 'Array', proto: true, forced: SLOPPY_METHOD || !USES_TO_LENGTH }, {
   every: function every(callbackfn /* , thisArg */) {
     return $every(this, callbackfn, arguments.length > 1 ? arguments[1] : undefined);
   }