+++ packages/core-js/modules/es.array.reduce-right.js
@@ -2,10 +2,14 @@
 var $ = require('../internals/export');
 var $reduceRight = require('../internals/array-reduce').right;
 var sloppyArrayMethod = require('../internals/sloppy-array-method');
+var arrayMethodUsesToLength = require('../internals/array-method-uses-to-length');
+
+var SLOPPY_METHOD = sloppyArrayMethod('reduceRight');
+var USES_TO_LENGTH = arrayMethodUsesToLength('reduceRight');
 
 // `Array.prototype.reduceRight` method
 // https://tc39.github.io/ecma262/#sec-array.prototype.reduceright
-$({ target: 'Array', proto: true, forced: sloppyArrayMethod('reduceRight') }, {
+$({ target: 'Array', proto: true, forced: SLOPPY_METHOD || !USES_TO_LENGTH }, {
   reduceRight: function reduceRight(callbackfn /* , initialValue */) {
     return $reduceRight(this, callbackfn, arguments.length, arguments.length > 1 ? arguments[1] : undefined);
   }