+++ packages/core-js/modules/es.array.some.js
@@ -2,10 +2,14 @@
 var $ = require('../internals/export');
 var $some = require('../internals/array-iteration').some;
 var sloppyArrayMethod = require('../internals/sloppy-array-method');
+var arrayMethodUsesToLength = require('../internals/array-method-uses-to-length');
+
+var SLOPPY_METHOD = sloppyArrayMethod('some');
+var USES_TO_LENGTH = arrayMethodUsesToLength('some');
 
 // `Array.prototype.some` method
 // https://tc39.github.io/ecma262/#sec-array.prototype.some
-$({ target: 'Array', proto: true, forced: sloppyArrayMethod('some') }, {
+$({ target: 'Array', proto: true, forced: SLOPPY_METHOD || !USES_TO_LENGTH }, {
   some: function some(callbackfn /* , thisArg */) {
     return $some(this, callbackfn, arguments.length > 1 ? arguments[1] : undefined);
   }