+++ tests/tests/es.array.reduce.js
@@ -1,4 +1,4 @@
-import { NATIVE, STRICT } from '../helpers/constants';
+import { STRICT } from '../helpers/constants';
 
 QUnit.test('Array#reduce', assert => {
   const { reduce } = Array.prototype;
@@ -40,12 +40,10 @@ QUnit.test('Array#reduce', assert => {
     assert.throws(() => reduce.call(null, () => { /* empty */ }, 1), TypeError);
     assert.throws(() => reduce.call(undefined, () => { /* empty */ }, 1), TypeError);
   }
-  if (NATIVE) {
-    assert.notThrows(() => reduce.call({
-      length: -1,
-      0: 1,
-    }, () => {
-      throw new Error();
-    }, 1) === undefined, 'uses ToLength');
-  }
+  assert.notThrows(() => reduce.call({
+    length: -1,
+    0: 1,
+  }, () => {
+    throw new Error();
+  }, 1) === 1, 'uses ToLength');
 });