+++ packages/core-js/modules/es.array.includes.js
@@ -4,7 +4,7 @@ var $includes = require('../internals/array-includes').includes;
 var addToUnscopables = require('../internals/add-to-unscopables');
 var arrayMethodUsesToLength = require('../internals/array-method-uses-to-length');
 
-var USES_TO_LENGTH = arrayMethodUsesToLength('indexOf', 0, true);
+var USES_TO_LENGTH = arrayMethodUsesToLength('indexOf', { ACCESSORS: true, 1: 0 });
 
 // `Array.prototype.includes` method
 // https://tc39.github.io/ecma262/#sec-array.prototype.includes