+++ tests/tests/es.array.splice.js
@@ -1,4 +1,4 @@
-import { STRICT } from '../helpers/constants';
+import { DESCRIPTORS, STRICT } from '../helpers/constants';
 
 QUnit.test('Array#splice', assert => {
   const { splice } = Array.prototype;
@@ -35,4 +35,14 @@ QUnit.test('Array#splice', assert => {
     return { foo: 1 };
   } };
   assert.same(array.splice().foo, 1, '@@species');
+  if (DESCRIPTORS) {
+    assert.notThrows(() => splice.call(Object.defineProperty({
+      length: -1,
+    }, 0, {
+      enumerable: true,
+      get() {
+        throw new Error();
+      },
+    }), 0, 2).length === 0, 'uses ToLength');
+  }
 });