+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
 ##### Unreleased
 - Fixed early implementations of `Array#{ every, forEach, includes, indexOf, lastIndexOf, reduce, reduceRight, slice, some, splice }` for the usage of `ToLength`
-- Added `RegExp#exec` dependency to methods which depends on the correctness of logic of this method, [#741](https://github.com/zloirock/core-js/issues/741)
+- Added `RegExp#exec` dependency to methods which depends on the correctness of logic of this method (`3.6.0-3.6.1` issue), [#741](https://github.com/zloirock/core-js/issues/741)
 - Refactored some internals
 
 ##### 3.6.1 - 2019.12.25